+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- Allow `re get comments --reviewed-only` to be combined with dataset filters
+
 # v0.38.19
 - Resume on every error